modify a UserBOImpl class

+++ src/main/java/lk/ijse/orm_coursework/bo/custom/impl/UserBOImpl.java
@@ -2,62 +2,79 @@ package lk.ijse.orm_coursework.bo.custom.impl;
 
 import lk.ijse.orm_coursework.bo.custom.UserBO;
 import lk.ijse.orm_coursework.bo.exception.DuplicateException;
-import lk.ijse.orm_coursework.bo.exception.InUseException;
-import lk.ijse.orm_coursework.bo.util.EntityDTOConvertor;
+import lk.ijse.orm_coursework.bo.exception.NotFoundException;
+import lk.ijse.orm_coursework.bo.util.EntityDTOConverter;
 import lk.ijse.orm_coursework.dao.DAOFactory;
 import lk.ijse.orm_coursework.dao.DAOTypes;
 import lk.ijse.orm_coursework.dao.custom.UserDAO;
 import lk.ijse.orm_coursework.dto.UserDTO;
 import lk.ijse.orm_coursework.entity.User;
 
-import java.sql.SQLException;
+
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 public class UserBOImpl implements UserBO {
 
     private  final UserDAO userDAO = DAOFactory.getInstance().getDAO(DAOTypes.USER);
-    private final EntityDTOConvertor convertor = new EntityDTOConvertor();
+    private final EntityDTOConverter converter = new EntityDTOConverter();
 
     @Override
-    public List<UserDTO> getAllUser() throws SQLException {
-        List<User> users = userDAO.getAll();
-        List<UserDTO> userDTOs = new ArrayList<>();
-        for (User user : users) {
-//            users.add(convertor.getUserDTO(user));
-            userDTOs.add(convertor.getUserDTO(user));
+    public List<UserDTO> getAllUsers() throws Exception {
+        List<User> userList = userDAO.getAll();
+        List<UserDTO> userDTOList = new ArrayList<>();
+        for (User user : userList) {
+            userDTOList.add(converter.getUserDTO(user));
         }
-        return userDTOs;
+        return userDTOList;
     }
 
     @Override
-    public void saveUser(UserDTO dto) throws DuplicateException, Exception {
-        User user = convertor.getUser(dto);
-        userDAO.save(user);
+    public String getLastUserId() throws Exception {
+        return userDAO.getLastId();
     }
 
     @Override
-    public void updateUser(UserDTO dto) throws SQLException {
-        User user = convertor.getUser(dto);
-        userDAO.update(user);
+    public boolean saveUsers(UserDTO t) throws Exception {
+        Optional<User> user = userDAO.findById(t.getUserId());
+        if (user.isPresent()) {
+            throw new DuplicateException("User already exists");
+        }
+        return userDAO.save(converter.getUserEntity(t));
 
     }
 
     @Override
-    public boolean deleteUser(String id) throws InUseException, Exception {
+    public boolean updateUsers(UserDTO t) throws Exception {
+        Optional<User> user = userDAO.findById(t.getUserId());
+        if (user.isEmpty()) {
+            throw new NotFoundException("User not found");
+        }
+        return userDAO.update(converter.getUserEntity(t));
+    }
+
+    @Override
+    public boolean deleteUsers(String id) throws Exception {
+        Optional<User> user = userDAO.findById(id);
+        if (user.isEmpty()) {
+            throw new NotFoundException("User not found");
+        }
         return userDAO.delete(id);
     }
 
     @Override
-    public String getNextId() throws SQLException {
-        String lastId = userDAO.getLastId();
-        char tableChar = 'U';
-        if (lastId != null){
-            String lastIdNumberString = lastId.substring(1);
-            int lastIdNumber = Integer.parseInt(lastIdNumberString);
-            int nextIdNumber = lastIdNumber + 1;
-            return String.format(tableChar + "%03d", nextIdNumber);
+    public List<String> getAllUserIds() throws Exception {
+        return userDAO.getAllIds();
+    }
+
+    @Override
+    public Optional<UserDTO> findByUserId(String id) throws Exception {
+        Optional<User> user = userDAO.findById(id);
+        if (user.isPresent()) {
+            return Optional.of(converter.getUserDTO(user.get()));
+        } else {
+            return Optional.empty();
         }
-        return tableChar + "001";
     }
 }
